Upgrade to Scala Native 0.4.3 and perform several improvements * fix issues relative to TimsTOF data processing * implement precursor m/z calibration during MGF generation * fix schema of mzDB (useful for mzDB creation) * improve the parsing of mzML meta-data parsing * SBT version changed from 1.3.8 to 1.5.5

+++ core/native/src/main/scala/com/github/mzdb4s/io/reader/param/NativeParamTreeParserImpl.scala
@@ -4,6 +4,7 @@ import scala.collection.Seq
 import scala.scalanative.libc.string.strstr
 import scala.scalanative.libc.string.strlen
 import scala.scalanative.unsafe._
+import scala.scalanative.unsigned._
 
 import com.github.mzdb4s.db.model.params._
 import com.github.mzdb4s.db.model.params.param._
@@ -196,7 +197,7 @@ object NativeParamTreeParserImpl {
     var j = 0
     var i = attrLen + 1 // skip attribute string + one char (")
     var reachedEndOfContent = false
-    val attributeContentAsCStr: CString = stackalloc[CChar](bufferSize)
+    val attributeContentAsCStr: CString = stackalloc[CChar](bufferSize.toULong)
 
     while (i < chunkLen && !reachedEndOfContent) {
       val c = attributeChunk(i)
